requirements.txt
numpy==1.24.3
openpyxl==3.0.10
pandas==2.0.3
pyxlsb==1.0.10
PyYAML==6.0.1
scikit_learn==1.3.0
streamlit==1.36.0
streamlit_authenticator==0.3.2
